commit my code pawn and request loan nft

See merge request quannv_edsolabs/dfy-mobile!1099

+++ lib/domain/model/pawn/offer_sent/offer_sent_detail_crypto_model.dart
@@ -41,7 +41,7 @@ class OfferSentDetailCryptoModel {
 
   int? bcOfferId;
 
-  int? liquidationThreshold;
+  double? liquidationThreshold;
 
   OfferSentDetailCryptoModel({
     this.id,
